Cleanup imports
JIRA: RAIL-4069
