+++ libs/sdk-ui-ext/src/internal/components/pluggableVisualizations/headline/HeadlineDescriptor.ts
@@ -1,6 +1,10 @@
 // (C) 2021-2022 GoodData Corporation
 import { bucketMeasure, IInsight, IInsightDefinition } from "@gooddata/sdk-model";
 import { ISettings } from "@gooddata/sdk-backend-spi";
+import { IHeadlineProps } from "@gooddata/sdk-ui-charts";
+import { BucketNames } from "@gooddata/sdk-ui";
+import { insightFilters } from "@gooddata/sdk-model";
+
 import {
     IVisualizationSizeInfo,
     IVisualizationDescriptor,
@@ -9,15 +13,12 @@ import {
 import { IFluidLayoutDescriptor } from "../../../interfaces/LayoutDescriptor";
 import { PluggableHeadline } from "./PluggableHeadline";
 import { DASHBOARD_LAYOUT_DEFAULT_KPI_HEIGHT, MAX_VISUALIZATION_HEIGHT } from "../constants";
-import { getReactEmbeddingCodeGenerator } from "../../../utils/embeddingCodeGenerator";
 import {
     bucketConversion,
     getInsightToPropsConverter,
+    getReactEmbeddingCodeGenerator,
     insightConversion,
-} from "../../../utils/embeddingCodeGenerator/insightToPropsConverter";
-import { IHeadlineProps } from "@gooddata/sdk-ui-charts";
-import { BucketNames } from "@gooddata/sdk-ui";
-import { insightFilters } from "@gooddata/sdk-model";
+} from "../../../utils/embeddingCodeGenerator";
 
 const hasSecondaryMeasure = (insight: IInsightDefinition) =>
     insight.insight.buckets.filter((bucket) => bucket.items.length > 0).length > 1;
