+++ libs/sdk-ui-ext/src/internal/utils/embeddingCodeGenerator/tests/getReactEmbeddingCodeGenerator.test.ts
